exclude tests from race detector
